% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizations.R
\name{AppendNormalizationCoefs}
\alias{AppendNormalizationCoefs}
\title{AppendNormalizationCoefs: Appends a new normalization coefinients to rMSI object.}
\usage{
AppendNormalizationCoefs(img, normName, normCoefs)
}
\arguments{
\item{img}{the rMSI object to append normalization coeficients.}

\item{normName}{the given name of the normalization.}

\item{normCoefs}{the normalizations coeficients.}
}
\value{
a rMSI image containing the new normalization field.
}
\description{
AppendNormalizationCoefs: Appends a new normalization coefinients to rMSI object.
}

