% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_calibration.R
\name{applyMassCalibration}
\alias{applyMassCalibration}
\title{applyMassCalibration.}
\usage{
applyMassCalibration(massModel, massRAW)
}
\arguments{
\item{massModel}{a massModel calculated using the calcMassModel() method.}

\item{massRAW}{a numeric vector containing the raw mass axist to be calibrated.}
}
\value{
a numeric vector with the calibrated mass axis.
}
\description{
Calibrates a mass axis using an already calculated mass model.
}
\examples{
# Set  five target mass peaks as example
target_peaks <- c(110, 200, 500, 700, 750) 

# Create a synthetic spectrum containing the five example peaks
raw_mass <- seq(from = 100, to = 800, by = 0.05)
raw_intensity <- 0.4*dnorm(raw_mass, mean = target_peaks[1], sd = 0.1) +
  0.5*dnorm(raw_mass, mean = target_peaks[2], sd = 0.1) +
  0.8*dnorm(raw_mass, mean = target_peaks[3], sd = 0.1) +
  0.5*dnorm(raw_mass, mean = target_peaks[4], sd = 0.1) +
  0.3*dnorm(raw_mass, mean = target_peaks[5], sd = 0.1)

# Set the reference masses for the  five target peaks (theoretical mass values)
ref_peaks <- c(110.1, 200.05, 499.95, 700.1, 750.15)

# Calculate the calibration model using the linear and loess methods
linModel <- rMSI::calcMassModel(ref_mz = ref_peaks, target_mz = target_peaks, method = "linear")
loessModel <- rMSI::calcMassModel(ref_mz = ref_peaks, target_mz = target_peaks, method = "loess", CalSpan = 20)

# Apply the calibration models to get a calibrated spectrum
mass_linCal <- rMSI::applyMassCalibration(linModel, raw_mass)
mass_loessCal <- rMSI::applyMassCalibration(loessModel, raw_mass)

# Plot the synthetic raw spectrum overlaied with its calibrations
rMSI::plotSpectra( raw_mass, raw_intensity, col = "black", ref_mass = ref_peaks) #raw with mass references marked
rMSI::plotSpectra( mass_linCal, raw_intensity, col = "red") #linear calibration
rMSI::plotSpectra( mass_loessCal, raw_intensity, col = "blue") #loess calibration

}
