% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_calibration.R
\name{applyMassCalibrationImage}
\alias{applyMassCalibrationImage}
\title{applyMassCalibrationImage.}
\usage{
applyMassCalibrationImage(img, massModel)
}
\arguments{
\item{img}{A image in rMSI data format .}

\item{massModel}{a massModel calculated using the calcMassModel() method.}
}
\value{
a rMSI img object with the mass calibration applied.
}
\description{
Apply a mass calibration model to a complete image. The new mass axis will be calculated using the supplied calibration model (massModel).
Then, the original imzML file will be overwritten with the calibrated mass axis. In case of data in continuous mode the calibration process is very fast since a 
single mass axis is recalculated and overwritten. For imzML in processed mode the calibration process may take some time since a new mass axis must be recalculated
for each pixel in the data.
}
\seealso{
\code{\link{applyMassCalibration}} and \code{\link{calcMassModel}}
}
