% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{AverageSpectrum}
\alias{AverageSpectrum}
\title{Computes the average spectrum of a whole rMSI image.}
\usage{
AverageSpectrum(img)
}
\arguments{
\item{img}{the rMSI image object.}
}
\value{
A vector with the average spectrum intensities. Masses are the same as the rMSI object.
}
\description{
Computes the average spectrum of a whole rMSI image.
}

