% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{BaseSpectrum}
\alias{BaseSpectrum}
\title{BaseSpectrum.
Computes the base spectrum of a whole rMSI image where the intensity value
for each mass bin is calculated as the maxium of all mass channels.}
\usage{
BaseSpectrum(img)
}
\arguments{
\item{img}{the rMSI image object.}
}
\value{
A vector with the base spectrum intensities. Masses are the same as the rMSI object.
}
\description{
BaseSpectrum.
Computes the base spectrum of a whole rMSI image where the intensity value
for each mass bin is calculated as the maxium of all mass channels.
}
