% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakMatrixMethods.R
\name{buildImgIdVectorFromPeakMatrix}
\alias{buildImgIdVectorFromPeakMatrix}
\title{buildImgIdVectorFromPeakMatrix.}
\usage{
buildImgIdVectorFromPeakMatrix(pkMat)
}
\arguments{
\item{pkMat}{an rMSIproc peak matrix object.}
}
\value{
an integer vector with all ID's of each image in the pkMat object.
}
\description{
Builds a integer vector containing all rMSI objects ID's accroding the peak matrix row order.
The resulting ID vector can be used to locate a spectrum ID in a peak matrix of multiple data sets.
}
