% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{builRasterImageFromCols}
\alias{builRasterImageFromCols}
\title{Build a image slice from specified datacube columns.}
\usage{
builRasterImageFromCols(Img, IonIndex, IonCount, Normalization = NULL)
}
\arguments{
\item{Img}{the rMSI object where the data is stored.}

\item{IonIndex}{the starting ion index to construc the image.}

\item{IonCount}{the number of ions used to build the image.}

\item{Normalization}{optionally a vector of the normalization factor for each pixel.}
}
\value{
a matrix with the same size as image size containing the pixel values.
}
\description{
Builds a image from the selected columns in the rMSI object. The image is returned arranged in a matrix containing each pixel value.
}
