% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcMassAxisBinSize}
\alias{CalcMassAxisBinSize}
\title{CalcMassAxisBinSize.}
\usage{
CalcMassAxisBinSize(mass, intensity)
}
\arguments{
\item{mass}{the mass axis.}

\item{intensity}{the intensity of a given spectrum.}
}
\value{
the bin size of each m/z channel.
}
\description{
Calc the bin size of a mass axis at each mass channels using simple peak-picking information.
}
