% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_calibration.R
\name{calcMassModel}
\alias{calcMassModel}
\title{calcMassModel.}
\usage{
calcMassModel(ref_mz, target_mz, method = "loess", CalSpan = 0.75)
}
\arguments{
\item{ref_mz}{a vector of reference masses (for exaple the theorical gold peaks).}

\item{target_mz}{manually slected masses to be fittet to ref_masses (must be the same length than ref_mz).}

\item{method}{a string with the method used for interpolation, valid methods are loess and linear}

\item{CalSpan}{the span for the loess method (ignored in linear mode).}
}
\value{
the calibration model.
}
\description{
Calculate the mass calibration model from a list of target masses and a list of reference masses. 
Two methods are available: linear and loess. The linear model is more robust to calibration artifacts and do not exhibit oscillations.
But, the loess model can provide a higher mass accuracy when the span parameter is tuned properly. When using the loess model, try to settle with a span value as low as possible.
In general, it is advised to use linear model for dataset acquired with high mass accuracy instruments (FTICR, Orbitrap) and keep the loess model for lower mass accuracy instruments (TOF).
}
\seealso{
\code{\link{applyMassCalibration}} for a complete mass calibration example.
}
