% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_calibration.R
\name{CalibrationWindow}
\alias{CalibrationWindow}
\title{CalibrationWindow.}
\usage{
CalibrationWindow(
  mass,
  intensity,
  peak_win_size = 20,
  win_title = "",
  CalibrationSpan = 0.75
)
}
\arguments{
\item{mass}{The mass vector of spectrum to calibrate.}

\item{intensity}{The intensity vector of spectrum to calibrate.}

\item{CalibrationSpan}{the span of the loess method for calibration.}
}
\value{
a list containing the calibration model and a data frame with calibration data. NULL is returned if calibration is aborted.
}
\description{
CalibrationWindow.
}
