% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Ccreate_rMSIXBinData}
\alias{Ccreate_rMSIXBinData}
\title{Ccreate_rMSIXBinData.}
\usage{
Ccreate_rMSIXBinData(rMSIobj, number_of_threads)
}
\arguments{
\item{rMSIobj:}{an rMSI object prefilled with a parsed imzML.}

\item{number_of_threads:}{number of threads used for imgStream encoding.}
}
\value{
the rMSI object with rMSIXBin inforation completed.
}
\description{
creates new rMSIXBin files (.XrMSI and .BrMSI). Previous files will be deleted.
}
