% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_additional_graphics.R
\name{chemFormula2Expression}
\alias{chemFormula2Expression}
\title{chemFormula2Expression.}
\usage{
chemFormula2Expression(strFormula)
}
\arguments{
\item{strFormula}{a string containing a chemical formula.}
}
\value{
a formated R expression that must be used with parse().
}
\description{
Converts a string containing a chemical forumla to an R expression that allows plotting the formula with sub-indices properly.
}
