% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CimzMLBinReadIntensity}
\alias{CimzMLBinReadIntensity}
\title{CimzMLBinReadIntensity.}
\usage{
CimzMLBinReadIntensity(
  ibdFname,
  NPixels,
  N,
  offset,
  dataTypeString,
  continuous
)
}
\arguments{
\item{ibdFname:}{full path to the ibd file.}

\item{NPixels:}{Total number of pixels in the image.}

\item{N:}{number of elemetns (or data point to read).}

\item{offset:}{offset in bytes at which the reading operation is started.}

\item{continuous:}{true if imzML data is in continuous mode}
}
\description{
Reads a single mass axis from the imzML file.
}
