% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CimzMLBinWriteModifyMass}
\alias{CimzMLBinWriteModifyMass}
\title{A method to use the imzMLwriter in modify mode to allow direct modification of mass axes for the calibration
This function modifies data of an ibd file with the following params}
\usage{
CimzMLBinWriteModifyMass(
  ibdFname,
  NPixels,
  mz_dataTypeString,
  int_dataTypeString,
  continuous,
  mzNew,
  mzOffset
)
}
\arguments{
\item{ibdFname:}{full path to the ibd file.}

\item{NPixels:}{Total number of pixels in the image.}

\item{mz_dataTypeString:}{String to specify the data format used to encode m/z values.}

\item{int_dataTypeString:}{String to specify the data format used to encode intensity values.}

\item{continuous:}{true if imzML data is in continuous mode}

\item{mzNew:}{A vector with the m/z values. Must be the same length as the original imzML mass target massa axis.}

\item{mzOffset:}{offset in the ibd file of the target mass axis.}
}
\description{
A method to use the imzMLwriter in modify mode to allow direct modification of mass axes for the calibration
This function modifies data of an ibd file with the following params
}
