% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CimzMLReadPeakList}
\alias{CimzMLReadPeakList}
\title{Method to read a peak list from an imzML file. Processed mode is assumed.
testingimzMLBinRead}
\usage{
CimzMLReadPeakList(ibdFname, imzML_peakList_descriptor, PixelID)
}
\arguments{
\item{ibdFname:}{full path to the ibd file.}

\item{imzML_peakList_descriptor:}{imzML file description as it is returned by the CimzMLParse() function.}

\item{PixelID:}{the pixel ID to read a peak list.}
}
\description{
Method to read a peak list from an imzML file. Processed mode is assumed.
testingimzMLBinRead
}
