% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cload_imzMLSpectra}
\alias{Cload_imzMLSpectra}
\title{Cload_imzMLSpectra
Load spectra into a Matrix object interpolating to the common mass axis when necessary.}
\usage{
Cload_imzMLSpectra(
  ibdFname,
  continuous,
  mass,
  mz_dataTypeString,
  int_dataTypeString,
  offsets
)
}
\arguments{
\item{ibdFname:}{full path to the ibd file.}

\item{continuous:}{true if imzML data is in continuous mode}

\item{mass:}{data common mass axis.}

\item{mz_dataTypeString:}{mass encoding data type name.}

\item{int_dataTypeString:}{mass encoding data type name.}

\item{offsets:}{subset of the run data of the imzML parsed file.}
}
\description{
Cload_imzMLSpectra
Load spectra into a Matrix object interpolating to the common mass axis when necessary.
}
