% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cload_rMSIXBinData}
\alias{Cload_rMSIXBinData}
\title{Cload_rMSIXBinData.}
\usage{
Cload_rMSIXBinData(path, fname)
}
\arguments{
\item{path:}{full path to the .XrMSI file.}

\item{fname:}{file name of the .XrMSI file without the extension.}
}
\value{
the rMSI object with rMSIXBin inforation completed.
}
\description{
Loads the data from the rMSIXBin files (.XrMSI and .BrMSI).
This method is used to load a previously stored rMSIXBin file.
}
