% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{ConvertBin2rMSIimg}
\alias{ConvertBin2rMSIimg}
\title{ConvertBin2rMSIimg.}
\usage{
ConvertBin2rMSIimg(in_data_dir, out_img_tar_file = NULL)
}
\arguments{
\item{in_data_dir}{data dir where the bin image is located.}

\item{out_img_tar_file}{if not NULL the imported image will be also stored as a tar file.}
}
\value{
the rMSI image object.
}
\description{
ConvertBin2rMSIimg.
}

