% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{ConvertrMSIimg2Bin}
\alias{ConvertrMSIimg2Bin}
\title{ConvertrMSIimg2Bin.}
\usage{
ConvertrMSIimg2Bin(img, out_data_dir)
}
\arguments{
\item{img}{a rMSI image object to be converted.}

\item{out_data_dir}{the resulting output data directory.}
}
\value{
nothing.
}
\description{
ConvertrMSIimg2Bin.
}
