% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportBrukerXMASSImage.R
\name{CountImagesInBrukerXml}
\alias{CountImagesInBrukerXml}
\title{CountImagesInBrukerXml.}
\usage{
CountImagesInBrukerXml(xml_path)
}
\arguments{
\item{xml_path}{the full path where XML file is stored.}
}
\value{
an integer with the total number of images found in the XML file.
}
\description{
Reads a Bruker's XML file and counts the number of regions.
This method is targetd to count the number of images to import previously of reading real data.
}
