% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CparseBrukerXML}
\alias{CparseBrukerXML}
\title{ParseBrukerXML.}
\usage{
CparseBrukerXML(xml_path)
}
\arguments{
\item{xml_path}{the full path where XML file is stored.}
}
\value{
ROI pixel coordinates arranged in a named list.
}
\description{
Reads a Bruker's xml file exported using fleximaging.
A list is returned where each element in the list is named according the ROI name.
Each element in the list consists in a data.frame with the pixels XY coordinates inside each ROI.
}
