% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CPeakList2PeakMatrix}
\alias{CPeakList2PeakMatrix}
\title{CPeakList2PeakMatrix.}
\usage{
CPeakList2PeakMatrix(
  RpeakList,
  BinTolerance = 5,
  BinFilter = 0.1,
  BinToleranceUsingPPM = TRUE
)
}
\arguments{
\item{RpeakList}{R peak list.}

\item{BinFilter}{the peaks bins non detected in at least the BinFitler*TotalNumberOfPixels spectra will be deleted.}

\item{BinToleranceUsingPPM}{if True the peak binning tolerance is specified in ppm, if false the tolerance is set using scans.}

\item{the}{tolerance used to merge peaks to the same bin. It is recomanded to use the half of peak width in ppm units.}
}
\value{
peak matrix.
}
\description{
Convert's an R peak list into a peak matrix.
}
