% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{CreateEmptyImage}
\alias{CreateEmptyImage}
\title{Create an empty rMSI object with defined mass axis and size.}
\usage{
CreateEmptyImage(x_size, y_size, mass_axis, pixel_resolution,
  ramdisk_folder = getwd())
}
\arguments{
\item{x_size}{the number of pixel in X direction.}

\item{y_size}{the number of pixel in Y direction.}

\item{mass_axis}{the mass axis.}

\item{pixel_resolution}{defined pixel size in um.}

\item{ramdisk_folder}{where ramdisk will be stored.

Creates an empty rMSI object with the provided parameters. This method is usefull to implement importation of new data formats
and synthetic datasets to test and develops processing methods and tools.}
}
\value{
the created rMSI object
}
\description{
Create an empty rMSI object with defined mass axis and size.
}

