% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imzMLreader.R
\name{dataPointBinaryEncoding}
\alias{dataPointBinaryEncoding}
\title{dataPointBinaryEncoding}
\usage{
dataPointBinaryEncoding(str_dataType)
}
\arguments{
\item{str_dataType}{a string with the C-style data type: int, long, float and double.}
}
\value{
a list containing an integer with the number of bytes to read for each data point and the R data type.
}
\description{
Get the number of bytes and data type that must be readed in binary streams for each encoded data point.
}
