% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{DeleteRamdisk}
\alias{DeleteRamdisk}
\title{Remove an rMSI object ramdisk}
\usage{
DeleteRamdisk(img)
}
\arguments{
\item{img}{an rMSI object.

Removes a rMSI objecte ramdisk.}
}
\description{
Remove an rMSI object ramdisk
}

