% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectPeaks.R
\name{DetectPeaks}
\alias{DetectPeaks}
\title{DetectPeaks'}
\usage{
DetectPeaks(mass, intensity, SNR = 5, WinSize = 20, OverSampling = 10)
}
\arguments{
\item{mass}{a vector containing the mass axis.}

\item{intensity}{a vector containinf the spectrum intensities.}

\item{SNR}{the minimum signal to noise ratio of retained peaks}

\item{WinSize}{the used windows size for peak detection}

\item{OverSampling}{the used oversampling value for interpolating the peak shape and improve mass and area calculation.}
}
\value{
a list containing mass, intensity, SNR, area and the binSize arround peak fields of detected peaks.
}
\description{
DetectPeaks'
}
