% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DetectPeaks_C}
\alias{DetectPeaks_C}
\title{DetectPeaks_C.}
\usage{
DetectPeaks_C(mass, intensity, SNR = 5, WinSize = 20L, UpSampling = 10L)
}
\arguments{
\item{mass}{a NumericVector containing the mass axis of the spectrum.}

\item{intensity}{a NumericVector where peaks must be detected.}

\item{SNR}{Only peaks with an equal or higher SNR are retained.}

\item{WinSize}{The windows used to detect peaks and caculate noise.}

\item{UpSampling}{the oversampling used for acurate mass detection and area integration.}
}
\value{
a NumerixMatrix of 5 rows corresponding to: mass, intensity of the peak, SNR, area and binSize.
}
\description{
Detect peaks from a Rcpp::NumericVector object and returns data in a R matrix.
This method is only exported to be use by R function DetectPeaks which is an actual R function.
The returned peak positions follows C indexing style, this is starts with zero.
}
