% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{.controlled_loadAbort}
\alias{.controlled_loadAbort}
\title{.controlled_loadAbort.}
\usage{
.controlled_loadAbort(text, close_signal = NULL)
}
\arguments{
\item{text}{to be promt at R console}

\item{close_signal}{the function to call}
}
\description{
.controlled_loadAbort.
}
