% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.debug_imzMLBinWriterSequential}
\alias{.debug_imzMLBinWriterSequential}
\title{Testing the imzMLwriter in sequential mode
This function creates a new ibd file with the provided data descibed in the following params}
\usage{
.debug_imzMLBinWriterSequential(
  ibdFname,
  mz_dataTypeString,
  int_dataTypeString,
  uuid,
  mzArray,
  intArray
)
}
\arguments{
\item{ibdFname:}{full path to the ibd file.}

\item{mz_dataTypeString:}{String to specify the data format used to encode m/z values.}

\item{int_dataTypeString:}{String to specify the data format used to encode intensity values.}

\item{uuid:}{16 bytes long UUID.}

\item{mzArray:}{A matrix with the m/z values for all pixels. Each pixel corresponds to a row. If there is only one row data will be saved in continuous mode}

\item{intArray:}{A matrix with the intensity values for all pixels. Each pixel corresponds to a row so the number of pixels is extracted from here.}
}
\description{
Testing the imzMLwriter in sequential mode
This function creates a new ibd file with the provided data descibed in the following params
}
