% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imzMLwriter.R
\name{export_imzML}
\alias{export_imzML}
\title{export_imzML}
\usage{
export_imzML(img, save_path)
}
\arguments{
\item{img}{an rMSI object.}

\item{save_path}{the output path where the .imzML and .ibd files will be stored.}
}
\description{
Exports and rMSI object to an imzML formated file.
}
