% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{ffVector2Matrix}
\alias{ffVector2Matrix}
\title{ffVector2Matrix.}
\usage{
ffVector2Matrix(ff_object)
}
\arguments{
\item{ff_object}{the ff object.}
}
\value{
a numeric matrix
}
\description{
Ensures the returned ff object is always a numeric matrix. This solves the problem of new ff release retuning a numeric vector when matrix has only a single row.
}
