% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPixelsInROI.R
\name{filterPixelsInROI}
\alias{filterPixelsInROI}
\title{filterPixelsInROI}
\usage{
filterPixelsInROI(
  intensityMatrix,
  ROIs,
  topPercentage = 0.01,
  lowPercentage = 0.1,
  magnitude = "RMS",
  plotResults = FALSE
)
}
\arguments{
\item{intensityMatrix}{Intensity matrix where rows represent pixels and columns represent peaks}

\item{ROIs}{Vector with a code value for each pixel in the intensity matrix.}

\item{topPercentage}{Percentage of top intensity pixels to remove.}

\item{lowPercentage}{Percentage of lower intensity pixels to remove.}

\item{magnitude}{Magnitude to use in filtering. RMS and TIC are the options allowed.}

\item{plotResults}{Flag to plot the results of the filtering in a violin plot.}
}
\value{

}
\description{
filterPixelsInROI
}
