% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyCmeans.R
\name{fuzzyCmeans}
\alias{fuzzyCmeans}
\title{fuzzyCmeans}
\usage{
fuzzyCmeans(
  intensityMatrix,
  numberOfClusters = 2,
  ROIs = NULL,
  m = 2,
  maxIterations = 100,
  epsilon = 1e-04
)
}
\arguments{
\item{intensityMatrix}{Intensity matrix where rows represent pixels and columns represent peaks.}

\item{numberOfClusters}{Number of clusters you want the data to be clusterized into. If ROIs is used, this parameter is effectless.}

\item{ROIs}{Vector with a code value for each pixel in the intensity matrix.}

\item{m}{Fuzzifier. When m = 1, the algorithm behaves like a normal k-means, as m increases, the partitions will be more and more diffuse. Recommended value is 2.}

\item{maxIterations}{Maximum number of iterations allowed during the execution.}

\item{epsilon}{Minimum step required between iterations to stop the algorithm. Values shoul be less than 1 and bigger than 0.}
}
\value{
The membership matrix, the centroids matrix and the hard partitioning clustering details.
}
\description{
fuzzyCmeans
}
