% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{getCoordsFromIds}
\alias{getCoordsFromIds}
\title{getCoordsFromIds Obtain images coords from a set of Ids.}
\usage{
getCoordsFromIds(Img, Ids)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Ids}{from which coords will be obtained.}
}
\value{
a matrix containing the coords.
}
\description{
getCoordsFromIds Obtain images coords from a set of Ids.
}

