% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{getCubeRowFromCoords}
\alias{getCubeRowFromCoords}
\title{Obtain the cube index and cube row of a given image coords.}
\usage{
getCubeRowFromCoords(Img, Coords)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Coords}{a coordinates vector of spectra to load represented as complex numbers where real part corresponds to X and imaginary to Y.}
}
\value{
a list of two vectors of cube index and row index in ff data objects list corresponding to given coords spectra.

Calculates where is located the data provided by pixel coordinates. The cube index and the cube row are returned.
}
\description{
Obtain the cube index and cube row of a given image coords.
}

