% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{getIdsFromMotorCoords}
\alias{getIdsFromMotorCoords}
\title{Obtain the image Identifiers from a given set of images motor coords}
\usage{
getIdsFromMotorCoords(Img, MotorCoords)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{MotorCoords}{a coords vector of spectra to load represented as complex numbers where real part corresponds to X and imaginary to Y.}
}
\value{
a vector of identifiers corresponding to given coords.
}
\description{
Calculate the pixel identifiers from the pixel motor coordinates.
}
