% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{getImageColsFromMass}
\alias{getImageColsFromMass}
\title{Obtain the rMSI datacube columns corresponding to a mass.}
\usage{
getImageColsFromMass(Img, Mass, Tolerance)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Mass}{the slected mass.}

\item{Tolerance}{a tolerance expressed in daltons around the slected mass.}
}
\value{
A list with: a Cols vector containing columns of the datacube, Mass and Tolerance recalculated to correctly fit the dataset.
}
\description{
Obtains the rMSI datacube columns that corresponds to a mass range defined by a central mass and a tolerance.
The central mass and tolerance are also recalculated to fit acurately in the dataset mass axis.
}
