% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakMatrixMethods.R
\name{getImgIdsFromPeakMatrixRows}
\alias{getImgIdsFromPeakMatrixRows}
\title{getImgIdsFromPeakMatrixRows.}
\usage{
getImgIdsFromPeakMatrixRows(pkMat, rows)
}
\arguments{
\item{pkMat}{an rMSIproc peak matrix object.}

\item{rows}{a vector of peak matrix rows.}
}
\value{
a list with the rMSI object ID's and image that correpond the specified rows.
}
\description{
Calculate the rMSI object ID's corresponding to a subset of row of a rMSIproc peak matrix.
}
