% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakMatrixMethods.R
\name{getPeakMatrixRowsFromImgIds}
\alias{getPeakMatrixRowsFromImgIds}
\title{getPeakMatrixRowsFromImgIds.}
\usage{
getPeakMatrixRowsFromImgIds(pkMat, img_num, ids)
}
\arguments{
\item{pkMat}{an rMSIproc peak matrix object.}

\item{img_num}{the number of the data set object to select in pkMat.}

\item{ids}{a vector of rMSI obj ID's.}
}
\value{
a vector containing the rows of peak matrix that correspond to the selected rMSI object ID's.
}
\description{
Obtains a vector of rMSIproc peak matrix rows corresponding to a vector of rMSI obj ID's.
}
