% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopIonsFromGroupComparison.R
\name{getTopIonsFromGroupComparison}
\alias{getTopIonsFromGroupComparison}
\title{getTopIonsFromGroupComparison}
\usage{
getTopIonsFromGroupComparison(
  groupsComparison,
  massVector,
  thresh_p = 0.05,
  thresh_FC = 1,
  maxNumberOfTopIons = 10,
  pAdjust = F,
  pAdjustMethod = "fdr"
)
}
\arguments{
\item{massVector}{}

\item{thresh_p}{}

\item{thresh_FC}{}

\item{maxNumberOfTopIons}{}

\item{pAdjust}{}

\item{pAdjustMethod}{}
}
\value{

}
\description{
getTopIonsFromGroupComparison
}
