% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/groupComparisonByPixelsClosestToROIsMedianSpectra.R
\name{groupComparisonByPixelsClosestToROIsMedianSpectra}
\alias{groupComparisonByPixelsClosestToROIsMedianSpectra}
\title{groupComparisonByPixelsClosestToROIsMedianSpectra}
\usage{
groupComparisonByPixelsClosestToROIsMedianSpectra(
  intensityMatrix,
  ROIs,
  numberOfPixels,
  testMethod = "Kruskal"
)
}
\arguments{
\item{intensityMatrix}{Intensity matrix where rows represent pixels and columns represent peaks.}

\item{ROIs}{Vector with a code value for each pixel in the intensity matrix.}

\item{numberOfPixels}{Number of pixels closest to the median specrum of each ROI to use in the comparison.}

\item{testMethod}{Test method to use in the group comparison. Options are: Wilcoxon, Students or Kruskal.}
}
\value{
List containing the p.value and FC for all ions at each group comparision.
}
\description{
groupComparisonByPixelsClosestToROIsMedianSpectra
}
