% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imzMLreader.R
\name{import_imzML}
\alias{import_imzML}
\title{import_imzML.}
\usage{
import_imzML(imzML_File, ibd_File = paste(sub("\\\\.[^.]*$", "", imzML_File),
  ".ibd", sep = ""), ramdisk_path, fun_progress = NULL, fun_text = NULL,
  close_signal = NULL)
}
\arguments{
\item{imzML_File}{full path to .imzML file.}

\item{ibd_File}{path to the binary file (default the same as imzML file but with .ibd extension)}

\item{ramdisk_path}{where the ramdisk will be created.}

\item{fun_progress}{This is a callback function to update the progress of loading data. See details for more information.}

\item{fun_text}{This is a callback function to update the label widget of loading data. See details for more information.}

\item{close_signal}{function to be called if loading process is abored.

 Imports an imzML image to an rMSI data object.
 It is recomanded to use rMSI::LoadMsiData directly instead of this function.}
}
\value{
an rMSI data object.
}
\description{
import_imzML.
}
