% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportBrukerXMASSImage.R
\name{importBrukerXmassImg}
\alias{importBrukerXmassImg}
\title{Import a MSI dataset from Bruker XMASS folder and a XML file.}
\usage{
importBrukerXmassImg(raw_data_folder, xml_file, ff_data_folder,
  max_ff_file_size_MB = 50)
}
\arguments{
\item{raw_data_folder}{Where the Bruker XMASS data is located.}

\item{xml_file}{Full path to the XML file.}

\item{ff_data_folder}{Full path to where the ff ramdisk will be stored.}

\item{max_ff_file_size_MB}{Max size in MB of each created ff file of the ramdisk.}
}
\value{
A rMSI data object pointing to the ramdisk.
}
\description{
\code{ importBrukerXmassImg } imports data using rMSI format and stores it to given ramdisk.
}
\details{
This function converts MSI data contained in Bruker XMASS directory into  rMSI data structure.
Only spectra contained in the XML file are imported. The ramdisk will be created at the specified
directory as a collection of ff files with the maximum specified size (50 MB is the default setting).
The rMSI object pointing to ramdisk and containing RAM stored variables is returned.
}

