% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imzMLreader.R
\name{imzMLparse}
\alias{imzMLparse}
\title{imzMLparse.
Parses a imzML file and obtain all relevant data for rMSI data object.}
\usage{
imzMLparse(fpath_xml, fpath_bin, fun_progress = NULL, close_signal = NULL)
}
\arguments{
\item{fpath_xml}{full path to imzML file (XML data).}

\item{fpath_bin}{full path to ibd file (binary data).}

\item{fun_progress}{callback to a pbar function.}

\item{close_signal}{function to call if error.}
}
\value{
a named list containing all relevant information for rMSI data object.
UUID: A string containing the UUID to be verified also in binary file.
continuous_mode: boolean indicating if binary data is in continuous mode (TRUE) or processed mode (FALSE).
compression_mz: boolean indicating if binary mz data is compressed (TRUE) or not (FALSE).
compression_int: boolean indicating if binary intensity data is compressed (TRUE) or not (FALSE).
mz_dataType: a string respresenting the data type used to encode mz axis (int, long, float or double).
int_dataType: a string respresenting the data type used to encode intensity data (int, long, float or double).
pixel_size_um: per pixel resolution in micrometers.
run_data: a data frame containing information of pixel location in image ans offsets in binary data file.
}
\description{
imzMLparse.
Parses a imzML file and obtain all relevant data for rMSI data object.
}

