% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{insertRasterImageAtMass}
\alias{insertRasterImageAtMass}
\title{Inserts a image at specified mass of a rMSI object.}
\usage{
insertRasterImageAtMass(Img, Mass, Tolerance, raster_matrix)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Mass}{the slected mass.}

\item{Tolerance}{a tolerance expressed in daltons around the slected mass.}

\item{raster_matrix}{a raster image represented as a matrix with pixel values.}
}
\description{
A raster image provided as a matrix is inserted at given Mass with a gaussian shape in the given Tolerance.
}
