% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceSpectrum.R
\name{InternalReferenceSpectrum}
\alias{InternalReferenceSpectrum}
\title{InternalReferenceSpectrum.}
\usage{
InternalReferenceSpectrum(img, reference)
}
\arguments{
\item{img}{MSI image to calculate internal reference spectrum.}

\item{reference}{the spectrum to use as reference for correlations.}
}
\value{
a list with the intensity vector corresponding to the reference spectrum, the score and the pixel ID selected as reference.
}
\description{
Calculates the dataset reference spectrum to use in label free alignment.
The reference spectrum is the spectrum in the dataset with the best correlation to average spectrum and with high TIC.
}
