% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{loadImageSliceFromCols}
\alias{loadImageSliceFromCols}
\title{Load a image slice from a specified datacube columns.}
\usage{
loadImageSliceFromCols(Img, Cols)
}
\arguments{
\item{Img}{the rMSI object where the data is stored.}

\item{Cols}{the columns indexes from which data will be taken.}
}
\value{
a data matrix containing the image slice.
}
\description{
Loads a slice of an rMSI object into RAM. The slice is determined by the columns specified in Cols parameter.
The returned value is not a MS image, in order to obtain a plotable image the function builRasterImageFromCols must be used.
}
