% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{loadImageSliceFromMass}
\alias{loadImageSliceFromMass}
\title{Load a image slice from a specified mass and tolerance.}
\usage{
loadImageSliceFromMass(Img, Mass, Tolerance)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Mass}{the slected mass.}

\item{Tolerance}{a tolerance expressed in daltons around the slected mass.}
}
\value{
a list with data matrix containing the image slice, the used mass and tolerance.
}
\description{
Loads a slice of an rMSI object into RAM. The slice is determined by a mass and a tolerance.
The returned value is not a MS image, in order to obtain a plotable image the function builRasterImageFromMass must be used.
}

