% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{loadImgChunkFromCoords}
\alias{loadImgChunkFromCoords}
\title{Loads a part of the img data in RAM.}
\usage{
loadImgChunkFromCoords(Img, Coords)
}
\arguments{
\item{Img}{the rMSI object where the data is stored.}

\item{Coords}{a coordinates vector of spectra to load represented as complex numbers where real part corresponds to X and imaginary to Y.}
}
\value{
a matrix containing the loaded spectra.
}
\description{
Loads a part of a rMSI object in computer memory. The spectra to load is specified by its pixel coordinates (Coords).
}
