% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{loadImgChunkFromCube}
\alias{loadImgChunkFromCube}
\title{Loads a part of a ff data img in RAM.}
\usage{
loadImgChunkFromCube(Img, Cube)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Cube}{The datacube index to load.}
}
\value{
a matrix containing the loaded spectra.
}
\description{
This function loads a specified ff datacubes.
It loads full cubes so it should be the most eficient and fast way of loading data to RAM.
}

