% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{loadImgChunkFromIds}
\alias{loadImgChunkFromIds}
\title{Loads a part of a img data in RAM.}
\usage{
loadImgChunkFromIds(Img, Ids, MassAxis = Img$mass)
}
\arguments{
\item{Img}{the rMSI object where the data is stored.}

\item{Ids}{Identifiers of spectra to load.}

\item{MassAxis}{a mass axis used for data interpolation. Defaults to Img$mass but may be different.}
}
\value{
a matrix containing the loaded spectra.
}
\description{
Loads a part of a rMSI object in computer memory. The spectra to load is specified by its identifiers (Ids).
}
