% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{loadImgCunckFromIds}
\alias{loadImgCunckFromIds}
\title{Loads a part of a ff data img in RAM.}
\usage{
loadImgCunckFromIds(Img, Ids)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Ids}{Identifiers of spectra to load.}
}
\value{
a matrix containing the loaded spectra.
}
\description{
Loads a part of a rMSI object in computer memory. The spectra to load is specified by its identifiers (Ids).
}

