% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load2ImagesWindow.R
\name{LoadTwoMsImages}
\alias{LoadTwoMsImages}
\title{LoadTwoMsImages}
\usage{
LoadTwoMsImages()
}
\value{
a list with the loaded rMSI objects
}
\description{
A GUI to select up to two MS images to load in rMSI data format
}
