% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MergeMassAxis}
\alias{MergeMassAxis}
\title{MergeMassAxis.}
\usage{
MergeMassAxis(mz1, bins1, mz2, bins2)
}
\arguments{
\item{mz1}{the first mass axis to merge.}

\item{bins1}{the bins size for the first mass axis.}

\item{mz2}{the second mass axis to merge.}

\item{intensity2}{the spectral intensities corresponding to the second mass axis.}
}
\value{
a list containing the common mass axis that represents mz1 and mz1 accurately and a boolean indicating if and error was raised.
}
\description{
Merges two mass axis in a single one using an apropiate bin size.
The resulting mass axis will display a bin size equal to the minimum of two supplied vectors. 
The bin size must be supplied along each input mass axis.
The first mass axis (mz1) can be a zero-length vector.
}
