% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NoiseEstimationFFTExpWin}
\alias{NoiseEstimationFFTExpWin}
\title{NoiseEstimationFFTExpWin.}
\usage{
NoiseEstimationFFTExpWin(x, filWinSize = 40L)
}
\arguments{
\item{x}{an Rcpp::NumericVector containing the spectrum intensities.}

\item{filWinSize}{an integer specified the cosinus win size in samples.}
}
\value{
an Rcpp::NumericVector containing the estimated noise.
}
\description{
Estimate the noise of a spectrum using a FFT filter and a decay exponential window in frequency domain.
}
