% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NoiseEstimationFFTExpWinMat}
\alias{NoiseEstimationFFTExpWinMat}
\title{NoiseEstimationFFTExpWinMat.}
\usage{
NoiseEstimationFFTExpWinMat(x, filWinSize = 40L)
}
\arguments{
\item{x}{an Rcpp::NumericMatrix containing the spectra intensities. Each spectrum in a row.}

\item{filWinSize}{an integer specified the cosinus win size in samples.}
}
\value{
an Rcpp::NumericMatrix containing the estimated noise in a matrix where each spectrum is a row.
}
\description{
Estimate the noise of some spectra using a FFT filter and a decay exponential window in frequency domain.
}
