% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizations.R
\name{NormalizeByAcqDegradation}
\alias{NormalizeByAcqDegradation}
\title{NormalizeByAcqDegradation: Normalizes an rMSI image to compensate ionization source degradation during acquisition.}
\usage{
NormalizeByAcqDegradation(img, winSize = 0.1)
}
\arguments{
\item{img}{the rMSI object to normalize.}

\item{winSize}{the window size use for smoothing (0 to 1).}
}
\value{
a rMSI image containing the normalizations$AcqTic field.
}
\description{
NormalizeByAcqDegradation: Normalizes an rMSI image to compensate ionization source degradation during acquisition.
}
