% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizations.R
\name{NormalizeByAcqusitionDegradation}
\alias{NormalizeByAcqusitionDegradation}
\title{NormalizeByAcqusitionDegradation: Normalizes an rMSI image to compensate ionization source degradation.}
\usage{
NormalizeByAcqusitionDegradation(img, winSize = 0.1)
}
\arguments{
\item{img}{the rMSI object to normalize.}

\item{winSize}{the window size use for smoothing (0 to 1).}
}
\value{
a rMSI image containing the normalizations$AcqTic field.
}
\description{
NormalizeByAcqusitionDegradation: Normalizes an rMSI image to compensate ionization source degradation.
}

