% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizations.R
\name{NormalizeMAX}
\alias{NormalizeMAX}
\title{NormalizeMAX: Calculates the MAX normalizatin of each pixel as the maximum of all intensities.}
\usage{
NormalizeMAX(img, remove_empty_pixels = FALSE)
}
\arguments{
\item{img}{the rMSI image object.}

\item{remove_empty_pixels}{boolean detailing if pixels detected to not contain data must be removed from normalization (smaller than mean-*sd).}
}
\value{
a rMSI image containing the normalizations$MAX field or MAXne if remove_empty_pixels is true..
}
\description{
NormalizeMAX: Calculates the MAX normalizatin of each pixel as the maximum of all intensities.
}

