% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizations.R
\name{NormalizeTargetPeaks}
\alias{NormalizeTargetPeaks}
\title{NormalizeTargetPeaks: Calculates the normalization of each pixel as the sum of areas of specified peaks.
The spectra that not contains target mass peaks will be removed by normalizing them by Inf.}
\usage{
NormalizeTargetPeaks(img, mz_vector, mz_window_size = 0.1, norm_name = "SelPK")
}
\arguments{
\item{img}{the rMSI image object.}

\item{mz_vector}{a vector containing m/z values of peaks to be used for normalization.}

\item{mz_window_size}{specifies the with of peak gaussian. Set to 0.1 Da by default which is OK for reflector TOF MALDI instrument}

\item{norm_name}{a name to be used for the normalization. Default is set to "SelPK".}
}
\value{
a rMSI image containing the normalizations$<name> field
}
\description{
NormalizeTargetPeaks: Calculates the normalization of each pixel as the sum of areas of specified peaks.
The spectra that not contains target mass peaks will be removed by normalizing them by Inf.
}
