% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{NormalizeTIC}
\alias{NormalizeTIC}
\title{NormalizeTIC: Calculates the TIC normalizatin of each pixel as the sum of all intensities.}
\usage{
NormalizeTIC(img)
}
\arguments{
\item{img}{the rMSI image object.}
}
\value{
a rMSI image containing the normalizations$TIC field.
}
\description{
NormalizeTIC: Calculates the TIC normalizatin of each pixel as the sum of all intensities.
}

