% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIWindow.R
\name{OpenMSI}
\alias{OpenMSI}
\title{Open a MS image from Hdd directly allowing the user to chose it from a file dialog.}
\usage{
OpenMSI()
}
\value{
the rMSI object containing the MS image.
}
\description{
A file choser dialog is presented to allow the selection of a .tar file containing an rMSI data object.
If a ramdisk has been previously created it will be used to speed up the loading process.
The image will be presented using the GUI and it will be returned as rMSI object.
}
