% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{PlotClusterImage}
\alias{PlotClusterImage}
\title{PlotClusterImage.
Plot a segmentation image with the user-given clusters.}
\usage{
PlotClusterImage(posMat, clusters, rotate = 0, pixel_size_um = 100)
}
\arguments{
\item{posMat}{a two columns matrix where first column ara the x coodrinates of values and second column the y coordinates.}

\item{clusters}{a vector with integer number according the cluster of each pixel.}

\item{rotate}{rotation to apply.}

\item{pixel_size_um}{the pixel resolution in um.}
}
\value{
a vector with the used color for each cluster sorted according clustering numering in assending order.
}
\description{
PlotClusterImage.
Plot a segmentation image with the user-given clusters.
}
