% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuzzyClusteringImage.R
\name{plotFuzzyClusteringImageG}
\alias{plotFuzzyClusteringImageG}
\title{plotFuzzyClusteringImageG}
\usage{
plotFuzzyClusteringImageG(
  peakMatrix,
  clusters,
  membership,
  title = "",
  contrastAlpha = FALSE,
  plotMembershipMap = FALSE
)
}
\arguments{
\item{peakMatrix}{rMSIprocPeakMatrix object.}

\item{clusters}{Integer vector that indicates at which cluster belongs each pixel.}

\item{membership}{Numeric vector that contains the membership value of the pixels to the cluster they belong in clusters.}

\item{title}{Title for the plot if desider. By default no title.}

\item{contrastAlpha}{If TRUE maps the range of memberships to the full range of alpha (0 to 1), increasing the contrast between pixels with different memberships.}

\item{plotMembershipMap}{If TRUE plots the membership image and the hard clustering at different images}
}
\value{

}
\description{
plotFuzzyClusteringImageG
}
