% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_additional_graphics.R
\name{plotMassDriftG}
\alias{plotMassDriftG}
\title{plotMassDriftG.}
\usage{
plotMassDriftG(
  peakMatrix,
  peakList,
  target_mass,
  error_range_ppm,
  min_SNR = 10,
  mass_offset = 0,
  title_label = "",
  normalization = NA,
  visible_legend = T,
  legend_title = "",
  N = 1
)
}
\arguments{
\item{peakMatrix}{an rMSIproc peak matrix.}

\item{peakList}{an rMSIproc peak list (no binning here!).}

\item{target_mass}{the target mass to represent.}

\item{error_range_ppm}{an error range in ppm to be represented around the target mass.}

\item{min_SNR}{the minimum signal to noise ratio to be represented.}

\item{mass_offset}{a mass offset to shift the plot mass axis relative to the target mass.}

\item{title_label}{a string to be used as plot title. Chemical forumlas will be parsed to produce better results.}

\item{normalization}{a vector containing the normalization value for each pixel or NA if no normalization should be applied.}

\item{visible_legend}{a boolean specfing if a legend detailing the included MS images must be displayed.}

\item{legend_title}{the title for the legend.}

\item{N}{numbe of dots to display for each pixel, only the N highest SNR will be displayed.}
}
\value{
a ggplot2 object.
}
\description{
Plot the mass shift observed at a target mass.
}
