% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_additional_graphics.R
\name{plotMassDriftImageComparedG}
\alias{plotMassDriftImageComparedG}
\title{plotMassDriftImageComparedG.}
\usage{
plotMassDriftImageComparedG(
  peakMatrix_RAW,
  peakMatrix_ALNG,
  peaklist_RAW,
  peaklist_ALNG,
  mass,
  error_range_ppm = 400,
  mass_offset_RAW = 0,
  title_label = "",
  ion_map_plot_cols = 2,
  ion_map_plot_rows = 2,
  ion_map_plot_byrow = T,
  ion_map_plot_rotations = rep(0, length(peakMatrix_ALNG$names)),
  ion_map_plot_labels = peakMatrix_ALNG$names,
  ion_map_plot_margin = 40,
  ion_map_plot_mirror_X = rep(F, length(peakMatrix_ALNG$names)),
  ion_map_plot_mirror_Y = rep(F, length(peakMatrix_ALNG$names)),
  ion_map_fixed_aspect_ratio = F
)
}
\arguments{
\item{peakMatrix_RAW}{an rMSIproc peak matrix obtained without the alginment routine (RAW).}

\item{peakMatrix_ALNG}{an rMSIproc peak matrix obtained with the alginment routine (Aligned).}

\item{peaklist_RAW}{a peak list generated with rMSIproc without the alginment routine (RAW).}

\item{peaklist_ALNG}{a peak list generated with rMSIproc with the alginment routine (Aligned).}

\item{mass}{the ion mass to be plot.}

\item{error_range_ppm}{a mass range to be plot specified in ppm.}

\item{mass_offset_RAW}{a mass offset applied to the RAW data to manually compensate possible mass shifts.}

\item{title_label}{the main title of the plot.}

\item{ion_map_plot_cols}{number of columns to arrange multiple images in the plotting area.}

\item{ion_map_plot_rows}{number of rows to arrange multiple images in the plotting area.}

\item{ion_map_plot_byrow}{a boolean idicating if the plotted images must be sorted by rows.}

\item{ion_map_plot_rotations}{a vector with the rotation in degree to apply to each image.}

\item{ion_map_plot_labels}{text labels to be used for each image.}

\item{ion_map_plot_margin}{a numeric value that determines the separation between images.}

\item{ion_map_plot_mirror_X}{a vector of booleans idicatinc if each image must be flipped horizontally.}

\item{ion_map_plot_mirror_Y}{a vector of booleans idicatinc if each image must be flipped vertically.}

\item{ion_map_fixed_aspect_ratio}{set this flag to true to fix the aspect ratio of the ion images.}
}
\value{
a ggplot2 object.
}
\description{
plots the image map displaying the mass shift at each raster position of two datasets.
this allows to compare the results of the alignment routine (raw vs. aliged).
}
