% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_additional_graphics.R
\name{plotMassDriftImageG}
\alias{plotMassDriftImageG}
\title{plotMassDriftImageG.}
\usage{
plotMassDriftImageG(
  peakMatrix,
  peakList,
  target_mass,
  error_range_ppm,
  plot_rows = 2,
  plot_cols = 2,
  plot_byrow = T,
  plot_rotations = rep(0, length(peakMatrix$names)),
  plot_mirror_X = rep(F, length(peakMatrix$names)),
  plot_mirror_Y = rep(F, length(peakMatrix$names)),
  plot_margin = 40,
  plot_labels = peakMatrix$names,
  title_label = "",
  fixed_aspect_ratio = F
)
}
\arguments{
\item{peakMatrix}{an rMSIproc peak matrix.}

\item{peakList}{an rMSIproc peak list (no binning here!).}

\item{target_mass}{the target mass to represent.}

\item{error_range_ppm}{an error range in ppm to be represented around the target mass.}

\item{plot_rows}{number of rows to arrange multiple images in the plotting area.}

\item{plot_cols}{number of columns to arrange multiple images in the plotting area.}

\item{plot_byrow}{a boolean idicating if the plotted images must be sorted by rows.}

\item{plot_rotations}{a vector with the rotation in degree to apply to each image.}

\item{plot_mirror_X}{a vector of booleans idicatinc if each image must be flipped horizontally.}

\item{plot_mirror_Y}{a vector of booleans idicatinc if each image must be flipped vertically.}

\item{plot_margin}{a numeric value that determines the separation between images.}

\item{plot_labels}{text labels to be used for each image.}

\item{title_label}{Text label for the plot main title.}

\item{fixed_aspect_ratio}{set this flag to true to fix the aspect ratio of the ion images.}
}
\value{
a ggplot2 object.
}
\description{
plots an image map displaying the mass shift at each raster position.
}
