% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMassImageByPeak.R
\name{plotMassImageByPeak}
\alias{plotMassImageByPeak}
\title{Plot a mass image of a up to 3 selected ions.}
\usage{
plotMassImageByPeak(img, mass.peak, tolerance = 0.25, XResLevel = 3,
  NormalizationCoefs = NULL, rotation = 0, show_axes = F,
  scale_to_global_intensity = F, vlight = 3, crop_area = NULL,
  intensity_limit = NULL)
}
\arguments{
\item{img}{an rMSI data object.}

\item{mass.peak}{a vector of up to 3 elements containing the mass of ions to plot.}

\item{tolerance}{a vector of up to 3 elements containing the mass range to plot for each ion.}

\item{XResLevel}{the interpolation factor (default is 3).}

\item{NormalizationCoefs}{optional parameter with a matrix containing the scaling factors for each pixel.}

\item{rotation}{the rotation of image expressed in deg. Valid values are: 0, 90, 180 and 270.}

\item{show_axes}{if true axis will be plotted. Otherwise a um scale and xy arrows  are drawn.}

\item{scale_to_global_intensity}{scale the image intensity to fit into the global intensity (only for RGB).}

\item{vlight}{the lighting of the plotted image.}

\item{crop_area}{the region of image to show in a vector formated as c(left, right, bottom, top) if null the whole image is plot.}

\item{intensity_limit}{limit the all pixels intensities to a given value. If null no limiting is used.

Plots a mass image. If only one mass ion are used a rainbow color code image is generated. If more ions are used each mass will
be encoded in an RGB color channel.}
}
\description{
Plot a mass image of a up to 3 selected ions.
}
