% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{plotParamAcqOrdered}
\alias{plotParamAcqOrdered}
\title{plotParamAcqOrdered.}
\usage{
plotParamAcqOrdered(img, Param, yAxisLabel = "Param")
}
\arguments{
\item{img}{rMSI object from wich data must be ploted.}

\item{Param}{a vector of elements to plot.}

\item{yAxisLabel.}{Param will be ploted ordered according the order of pixels in MALDI acqusition.}
}
\description{
plotParamAcqOrdered.
}
