% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeakMatrix.R
\name{plotPeakImage}
\alias{plotPeakImage}
\title{plotPeakImage.}
\usage{
plotPeakImage(
  peakMatrix,
  mz = NULL,
  column = NULL,
  matrix = "intensity",
  normalization = NULL,
  nrow = 2,
  ncol = 2,
  byrow = T,
  margin = 20,
  img_names = peakMatrix$names,
  labels = img_names,
  rotations = rep(0, length(img_names)),
  mirror_x = rep(F, length(img_names)),
  mirror_y = rep(F, length(img_names)),
  pixel_size_um = 100,
  light = 8
)
}
\arguments{
\item{peakMatrix}{the peak matrix in an rMSIproc object.}

\item{mz}{the peak mass to plot, the nearest peak mass will be ploted.}

\item{column}{the column of the peak matrix to plot.}

\item{matrix}{the name of the peak matrix to plot.}

\item{normalization}{the name of normalization to use.}

\item{nrow}{number of rows of the plotted matrix layout (only used if byrow == F)}

\item{ncol}{number of cols of the plotted matrix layout (only used if byrow == T)}

\item{byrow}{a bool specifing if images must be arranged in rows.}

\item{margin}{the separation between plotted images.}

\item{img_names}{a character vector with img names in the desierd plotting order.}

\item{labels}{an alternative character vector with the labels to be displayed for each image.}

\item{rotations}{a vector with the rotation of each images specified in degrees.}

\item{mirror_x}{a bool vector specifing if each image must be flipped or not in X direction prior to rotation.}

\item{mirror_y}{a bool vector specifing if each image must be flipped or not in Y direction prior to rotation.}

\item{pixel_size_um}{the pixel resolution in um.}

\item{light}{the lighting of the plotted image.}
}
\description{
plot the ion image map of a given mass or column of a rMSIproc peak matrix object.
If the peak matrix contains data from various datasets the images will be layout horizontally or vertically.
At leas mz or column must be specified.
}
