% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_additional_graphics.R
\name{plotPeakImageG}
\alias{plotPeakImageG}
\title{plotPeakImageG.}
\usage{
plotPeakImageG(
  peakMatrix,
  mass,
  normalization = NA,
  plot_rows = 2,
  plot_cols = 2,
  plot_byrow = T,
  plot_rotations = rep(0, length(peakMatrix$names)),
  plot_mirror_X = rep(F, length(peakMatrix$names)),
  plot_mirror_Y = rep(F, length(peakMatrix$names)),
  plot_margin = 40,
  plot_labels = peakMatrix$names,
  title_label = "",
  fixed_aspect_ratio = F,
  display_colorbar = T
)
}
\arguments{
\item{peakMatrix}{an rMSIproc peak matrix.}

\item{mass}{the ion mass to be plot.}

\item{normalization}{a vector containing the normalization value for each pixel or NA if no normalization should be applied.}

\item{plot_rows}{number of rows of the plotted matrix layout (only used if byrow == F).}

\item{plot_cols}{number of cols of the plotted matrix layout (only used if byrow == T).}

\item{plot_byrow}{a bool specifing if images must be arranged in rows.}

\item{plot_rotations}{a vector with the rotation of each images specified in degrees.}

\item{plot_mirror_X}{a bool vector specifing if each image must be flipped or not in X direction prior to rotation.}

\item{plot_mirror_Y}{a bool vector specifing if each image must be flipped or not in Y direction prior to rotation.}

\item{plot_margin}{the separation between plotted images.}

\item{plot_labels}{an alternative character vector with the labels to be displayed for each image.}

\item{title_label}{the main title of the plot}

\item{fixed_aspect_ratio}{set this flag to true to fix the aspect ratio of the ion images.}

\item{display_colorbar}{set if the colour bar must be displayed.}
}
\value{
a ggplot2 object.
}
\description{
plots and ion map image using the rMSIproc peak matrix.
This function is equivalent to the rMSIproc::plotPeakImage but used ggplot to produce more publication-ready results.
}
