% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMassImageByPeak.R
\name{plotRGBDataOnImg}
\alias{plotRGBDataOnImg}
\title{plotRGBDataOnImg: Function to plot generic data on a RGB raster using rMSI pixel locations.}
\usage{
plotRGBDataOnImg(img, Rvalues, Gvalues, Bvalues, RText, GText, BText,
  XResLevel = 3, light = 3, rotation = 0)
}
\arguments{
\item{img}{an rMSI object.}

\item{Rvalues}{a vector of values to be ploted as red layer sorted according IDs of rMSI object.}

\item{Gvalues}{a vector of values to be ploted as green layer sorted according IDs of rMSI object.}

\item{Bvalues}{a vector of values to be ploted as blue layer sorted according IDs of rMSI object.}

\item{RText}{Text to plot as Red scale label.}

\item{GText}{Text to plot as Green scale label.}

\item{BText}{Text to plot as Blue scale label.}

\item{XResLevel}{the interpolation to use in plot.}

\item{light}{the light to aplied to raster.}

\item{rotation}{image rotation in degree.}
}
\description{
plotRGBDataOnImg: Function to plot generic data on a RGB raster using rMSI pixel locations.
}

