% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectraWidget.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Plot Mass Spectra in a interactive way.}
\usage{
plotSpectra(
  mass = NULL,
  intensity = NULL,
  peaks_mass = NULL,
  peaks_intensity = NULL,
  ref_mass = NULL,
  col = ""
)
}
\arguments{
\item{mass}{the mass axis of spectrum}

\item{intensity}{a intensity vector of spectrum}

\item{peaks_mass}{a vector of peak masses to be labeled on spectrum}

\item{peaks_intensity}{a vector of peak intensities to be labeled on spectrum}

\item{ref_mass}{a vector of reference masses to be represented as vertical dashed lines}

\item{col}{the color for the spectrum

Simply call this function each time you want to plot a mass spectrum and it will be overlayed with the current spectra.
Peaks can be specified in the same call and it will be labeled over the spectrum.}
}
\description{
Plot Mass Spectra in a interactive way.
}
