% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopIonsInROIsComparaison.R
\name{plotTopIonsInROIsComparaison}
\alias{plotTopIonsInROIsComparaison}
\title{plotTopIonsInROIsComparaison}
\usage{
plotTopIonsInROIsComparaison(topIonsList, title)
}
\arguments{
\item{topIonsList}{List returned by the function "getTopIonsFromGroupComparison"}

\item{title}{Title of the plots}
}
\value{

}
\description{
plotTopIonsInROIsComparaison
}
