% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMassImageByPeak.R
\name{plotVariousMassImagesByPeak}
\alias{plotVariousMassImagesByPeak}
\title{Plots a image using various MS image objects and the same intensity scale for every MS image.}
\usage{
plotVariousMassImagesByPeak(..., mass.peak, tolerance = 0.25,
  NormalizationName = NULL, rotation = 0, flipV = F, flipH = F,
  light = 5)
}
\arguments{
\item{...}{MS images in rMSI format as various arguments to be plotted together.}

\item{mass.peak}{m/z selected to plot MS image}

\item{tolerance}{mass tolerance to represent the MS image.}

\item{NormalizationName}{a string representing the normalization to use in the plot.}

\item{rotation}{a vector of rotation in degree to perform on plot for each image.
If only one value is supplied, then all image will be rotated the same.}

\item{flipV}{vector of booleans indicating if images must be flipped vertically.}

\item{flipH}{vector of booleans indicating if images must be flipped horizontally.}

\item{light}{the lighting factor used in the plot.}
}
\description{
Plots a image using various MS image objects and the same intensity scale for every MS image.
}
