% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{ProcessImages}
\alias{ProcessImages}
\title{ProcessImages.}
\usage{
ProcessImages(
  proc_params,
  data_description,
  verifyImzMLChecksums = F,
  numOfThreads = min(parallel::detectCores()/2, 6),
  memoryPerThreadMB = 100
)
}
\arguments{
\item{proc_params}{a ProcParams object containing the processing parameters}

\item{data_description}{a DataInfo object containing the imzML files paths and parsed ROIs.}

\item{verifyImzMLChecksums}{a boolean indicating whether imzML checksums must be validated or not (default is not since it takes a long time).}

\item{numOfThreads}{the number number of threads used to process the data.}

\item{memoryPerThreadMB}{maximum allowed memory by each thread. The total number of trehad will be two times numOfThreads, so the total memory usage will be: 2*numOfThreads*memoryPerThreadMB.}
}
\value{

}
\description{
Process a single image or multiple images with the complete processing workflow.
}
