% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imzMLreader.R
\name{read_imzML_peaklists}
\alias{read_imzML_peaklists}
\title{read_imzML_peaklists.}
\usage{
read_imzML_peaklists(img)
}
\arguments{
\item{img}{a rMSIObj with valid imzML data in processed mode.}
}
\value{
a rMSIObj with the peak list in the img$data$peaklist field.
}
\description{
Reads the peak list in an imzML file containing data in processed mode.
Use this function carefully since all peak lists are loaded in memory.
}
