% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{ReadBrukerRoiXML}
\alias{ReadBrukerRoiXML}
\title{ReadBrukerRoiXML.}
\usage{
ReadBrukerRoiXML(img, xml_file)
}
\arguments{
\item{img}{an rMSI MS image object.}

\item{xml_file}{a full path to a Bruker XML ROI file.}
}
\value{
a list containing lists of ID's for each ROI.
}
\description{
Reads a Bruker ROI XML file and matches it to an rMSI image object.
A list of rMSI ID's contained in each Bruker's ROI will be returned.
}
