% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{readimzML_singlePixelPeakList}
\alias{readimzML_singlePixelPeakList}
\title{readimzML_singlePixelPeakList.}
\usage{
readimzML_singlePixelPeakList(imzML_file, pixel_id)
}
\arguments{
\item{imzML_file}{path to the imzML file containing the peak list (with or without extension).}

\item{pixel_id}{the pixel ID in the imzML to read.}
}
\value{
a list containg the peak information for the given pixel.
}
\description{
Reads a single pixel peaklist directely from an imzML file in processed mode.
Extra peak information will be available if the peak list was previously stored in the rMSI peak list format. 
The extra peak infor include peak area, SNR and bin size.
}
