% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{ROIAverageSpectra}
\alias{ROIAverageSpectra}
\title{ROIAverageSpectra.}
\usage{
ROIAverageSpectra(img, roi_list)
}
\arguments{
\item{img}{an rMSI object.}

\item{roi_list}{a roi list in the format returned by ReadBrukerRoiXML().}
}
\value{
all rois average spectra arranges in a list.
}
\description{
Calculates the average spectrum within each roi.
}
