% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{ROIAverageSpectraByIds}
\alias{ROIAverageSpectraByIds}
\title{ROIAverageSpectraByIds.}
\usage{
ROIAverageSpectraByIds(img, Ids)
}
\arguments{
\item{img}{an rMSI object.}

\item{Ids}{Identifiers of spectra to use for average calculation.}
}
\value{
the ROI average spectrum.
}
\description{
Calculates the average spectrum within each roi specified by a vector of pixel ID's
}
