% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{RunPreProcessing}
\alias{RunPreProcessing}
\title{RunPreProcessing}
\usage{
RunPreProcessing(
  proc_params,
  img_lst,
  verifyImzMLChecksums = F,
  numOfThreads = min(parallel::detectCores()/2, 6),
  memoryPerThreadMB = 200
)
}
\arguments{
\item{proc_params}{a ProcParams object containing the processing parameters}

\item{img_lst}{a rMSI objects lis to process.}

\item{numOfThreads}{the number number of threads used to process the data.}

\item{memoryPerThreadMB}{maximum allowed memory by each thread. The total number of trehad will be two times numOfThreads, so the total memory usage will be: 2*numOfThreads*memoryPerThreadMB.}
}
\value{

}
\description{
Process a single image or multiple images with the complete processing workflow.
}
