% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{saveImageSliceAtCols}
\alias{saveImageSliceAtCols}
\title{Overwrite an image slice of an rMSI object.}
\usage{
saveImageSliceAtCols(Img, Cols, dm)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Cols}{the columns indexes at which data will be overwriten.}

\item{dm}{the new data matrix overwrite current rMSI slected columns.}
}
\description{
Overwrite a part of a ramdisk in rMSI object with the provided data matrix in the specified columns.
}
