% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{saveImageSliceAtMass}
\alias{saveImageSliceAtMass}
\title{Overwrite an image slice of an rMSI object.}
\usage{
saveImageSliceAtMass(Img, Mass, dm)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Mass}{the central mass at which data will be overwriten.}

\item{dm}{the new data matrix overwrite current rMSI slected columns.}
}
\description{
Overwrite a part of a ramdisk in rMSI object with the provided data matrix in the specified mass range.
The mass range is obtained by the provided central mass and the necessary columns to fit the full data matrix in rMSI.
}
