% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{saveImgChunkAtCoords}
\alias{saveImgChunkAtCoords}
\title{Stores a data matrix to a part of ff data img.}
\usage{
saveImgChunkAtCoords(Img, Coords, dm)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Coords}{a coordinates vector of spectra to load represented as complex numbers where real part corresponds to X and imaginary to Y.}

\item{dm}{a matrix containing the data to overwrite the rMSI object.}
}
\description{
Overwrites a part of rMSI object with the provided data matrix. The part of rMSI object that is overwrited is specified by the Coords vector.
}
