% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{saveImgChunkAtIds}
\alias{saveImgChunkAtIds}
\title{Stores a data matrix to a part of ff data img.}
\usage{
saveImgChunkAtIds(Img, Ids, dm)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Ids}{Identifiers of spectra to be overwrited.}

\item{dm}{a matrix containing the data to overwrite the rMSI object.}
}
\description{
Overwrites a part of rMSI object with the provided data matrix. The part of rMSI object that is overwrited is specified by the Ids vector.
}

