% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{saveImgChunkToCube}
\alias{saveImgChunkToCube}
\title{Save a data matrix to a whole cube in rMSI object}
\usage{
saveImgChunkToCube(Img, Cube, dm)
}
\arguments{
\item{Img}{the rMSI object where the data will be overwrited (ramdisk).}

\item{Cube}{the cube index that will be overwrited}

\item{dm}{the data matrix that will overwrite the selected cube}
}
\description{
Save a data matrix to a whole cube in rMSI object
}

