% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{SaveMsiData}
\alias{SaveMsiData}
\title{Save a rMSI object to disk in a compressed .tar file.}
\usage{
SaveMsiData(imgData, data_file)
}
\arguments{
\item{imgData}{a rMSI objecte with a image with a working ramdisk.}

\item{data_file}{Full path to hdd location to save image as a tar file.}
}
\description{
Save a rMSI object to disk in a compressed .tar file.
}

