% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Smoothing_SavitzkyGolay}
\alias{Smoothing_SavitzkyGolay}
\title{Smoothing_SavitzkyGolay.}
\usage{
Smoothing_SavitzkyGolay(x, sgSize = 5L)
}
\arguments{
\item{x}{the data vector to smooth.}

\item{sgSize}{valid values are: 5, 7, 9, 11, 13, 15.}
}
\value{
the smoothed data vector.
}
\description{
Computes the Savitzky-Golay smoothing of a vector x using a filter size of sgSize.
}
