% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{SortIDsByAcquisition}
\alias{SortIDsByAcquisition}
\title{SortIDsByAcquisition: Order the rMSI pixel IDs according the acqusition sequence (first acquired pixel is the first one).}
\usage{
SortIDsByAcquisition(img)
}
\arguments{
\item{img}{a rMSI image.}
}
\value{
a vector of ID's ordered acording acquisiton.
}
\description{
SortIDsByAcquisition: Order the rMSI pixel IDs according the acqusition sequence (first acquired pixel is the first one).
}
