% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakMatrixMethods.R
\name{summary.rMSIprocPeakMatrix}
\alias{summary.rMSIprocPeakMatrix}
\title{Displays a summary of a peak matrix.}
\usage{
\method{summary}{rMSIprocPeakMatrix}(x)
}
\arguments{
\item{x}{rMSIproc peak matrix object.}
}
\description{
Displays a summary of a peak matrix.
}
