% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestHanningWindow}
\alias{TestHanningWindow}
\title{TestHanningWindow.}
\usage{
TestHanningWindow(mass, WinSize = 20L, UpSampling = 10L)
}
\arguments{
\item{mass}{a NumericVector containing the mass axis of the spectrum.}

\item{WinSize}{The windows used to detect peaks and caculate noise.}

\item{UpSampling}{the oversampling used for acurate mass detection and area integration.}
}
\value{
a NumericVector containing the Hanning Window.
}
\description{
Method to test the implementation of Hanning window in R session.
}
