% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestPeakInterpolation_C}
\alias{TestPeakInterpolation_C}
\title{TestPeakInterpolation_C.}
\usage{
TestPeakInterpolation_C(
  mass,
  intensity,
  peakIndex,
  WinSize = 20L,
  UpSampling = 10L,
  useHanning = FALSE,
  Iterations = 1L
)
}
\arguments{
\item{mass}{a NumericVector containing the mass axis of the spectrum.}

\item{intensity}{a NumericVector where peaks must be detected.}

\item{peakIndex}{the location of the peak to interpolate in the spectrum.}

\item{WinSize}{The windows used to detect peaks and caculate noise.}

\item{UpSampling}{the oversampling used for acurate mass detection and area integration.}

\item{useHanning}{if hanning windowing must be used befor interpolation.}

\item{Iterations}{number of iterations to perform. This is just for testing interpolation efficiency}
}
\value{
a NumerixVector with the FFT interpolated peak shape.
}
\description{
TestPeakInterpolation_C.
}
