% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{uuid_timebased}
\alias{uuid_timebased}
\title{uuid.}
\usage{
uuid_timebased()
}
\value{
the generated UUID encoded in a text string.
}
\description{
Generates a timecode-based 16-bytes UUID.
The generated bytes are generated using the following pattern:
bytes 0 and 1: Year
byte 2: Month
byte 3: Day
byte 4: Hour
byte 5: Minute
byte 6: Second
bytes 7 to 15: random.
}
