% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imzMLreader.R
\name{xmlParseSpectra}
\alias{xmlParseSpectra}
\title{xmlParseSpectra.}
\usage{
xmlParseSpectra(xmlRunNode)
}
\arguments{
\item{xmlRunNode}{an XML node containing the imzML run tree.}
}
\value{
a named data frame containing pixels positions in image (x, y) and spectra poisitions in binary file (offsets).
}
\description{
Extracts all important information to read the imzML binary file from the XML run node.
Internal use only.
}

