% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imzMLreader.R
\name{xmlParseSpectrum}
\alias{xmlParseSpectrum}
\title{xmlParseSpectrum.}
\usage{
xmlParseSpectrum(xmlSingleSpectrum, mzSRef, inSRef)
}
\arguments{
\item{xmlSingleSpectrum}{a xml node containing data for a single spectrum.}

\item{mzSRef}{a string with the naming reference used to search mz array.}

\item{inSRef}{a string with the naming reference used to search intensity array.}
}
\value{
a named list containing all parameters of the spectrum.
}
\description{
Parses an XML node in imzML format containing a single spectrum. All meaning data is returned as named list.
Internal use only.
}
