% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{builRasterImageFromCols}
\alias{builRasterImageFromCols}
\title{Build a image slice from specified datacube columns.}
\usage{
builRasterImageFromCols(Img, Cols, method = "max", Normalization = NULL)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Cols}{the columns indexes from which data will be taken.}

\item{method}{the method used to calculate the pixel value. Can be max or mean.}

\item{Normalization}{optionally a vector of the normalization factor for each pixel.}
}
\value{
a matrix with the same size as image size containing the pixel values.
}
\description{
Builds a image from the selected columns in the rMSI object. The image is returned arranged in a matrix containing each pixel value.
}

