% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIMethods.R
\name{CalibrateImage}
\alias{CalibrateImage}
\title{Apply a new mass axis to a complete image.
A GUI to calibrate the mean spectrum will be shown.}
\usage{
CalibrateImage(img, output_fname)
}
\arguments{
\item{A}{image in rMSI data format .}

\item{full}{path to store the output image.}
}
\description{
Apply a new mass axis to a complete image.
A GUI to calibrate the mean spectrum will be shown.
}

