% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIMethods.R
\name{CalibrateImage}
\alias{CalibrateImage}
\title{Apply a new mass axis to a complete image.
A GUI to calibrate the mean spectrum will be shown.}
\usage{
CalibrateImage(img, output_fname, newMzAxis = NULL)
}
\arguments{
\item{img}{A image in rMSI data format .}

\item{output_fname}{full path to store the output image.}

\item{newMzAxis}{if a new mz axis specified it is used for the calibration window}
}
\description{
Apply a new mass axis to a complete image.
A GUI to calibrate the mean spectrum will be shown.
}

