% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIMethods.R
\name{CalibrateImage}
\alias{CalibrateImage}
\title{Apply a new mass axis to a complete image.}
\usage{
CalibrateImage(img, new_mass, output_fname)
}
\arguments{
\item{A}{image in rMSI data format .}

\item{The}{new mass axis.}

\item{full}{path to store the output image.}
}
\description{
Apply a new mass axis to a complete image.
}

