% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrationWindow.R
\name{CalibrationWindow}
\alias{CalibrationWindow}
\title{CalibrationWindow.}
\usage{
CalibrationWindow(mass, intensity)
}
\arguments{
\item{mass}{The mass vector of spectrum to calibrate.}

\item{intensity}{The intensity vector of spectrum to calibrate.}
}
\value{
a the calibrated mass axis.
}
\description{
CalibrationWindow.
}

