% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cload_imzMLSpectra}
\alias{Cload_imzMLSpectra}
\title{Cload_imzMLSpectra
Load spectra into a Matrix object interpolating to the common mass axis when necessary.}
\usage{
Cload_imzMLSpectra(rMSIobj, pixelIDs)
}
\arguments{
\item{rMSIobj:}{an rMSI object prefilled with a parsed imzML.}

\item{pixelIDs:}{pixel ID's of the spectra to load in C-style indexing (starting at 0).}
}
\description{
Cload_imzMLSpectra
Load spectra into a Matrix object interpolating to the common mass axis when necessary.
}
