% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{CreateSubDataset}
\alias{CreateSubDataset}
\title{CreateSubDataset.}
\usage{
CreateSubDataset(img, id, ramdisk_path)
}
\arguments{
\item{img}{the original rMSI object.}

\item{id}{a vector of ID's to retain in the sub data set.}

\item{ramdisk_path}{a full disk path where the new ramdisk will be stored.}
}
\value{
the sub rMSI object.
}
\description{
Creates a new rMSI image object from sub-set of selected pixels by ID's.
}
