% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libimgramdisk.R
\name{getCubeRowFromIds}
\alias{getCubeRowFromIds}
\title{Obtain the cube index and cube row of a given image Ids.}
\usage{
getCubeRowFromIds(Img, Ids)
}
\arguments{
\item{Img}{the rMSI object where the data is stored (ramdisk).}

\item{Ids}{Identifiers of spectra.}
}
\value{
a list of two vectors of cube index and row index in ff data objects list corresponding to given coords spectra.
}
\description{
Calculates where is located the data provided by pixel Identifiers. The cube index and the cube row are returned.
}
