% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportBrukerXMASSImage.R
\name{importBrukerXmassImg}
\alias{importBrukerXmassImg}
\title{Import a MSI dataset from Bruker XMASS folder and a XML file.}
\usage{
importBrukerXmassImg(raw_data_full_path, resolution_um, xml_file_full_path,
  txt_spectrum_path = NULL, selected_img = 1, ...)
}
\arguments{
\item{raw_data_full_path}{Where the Bruker XMASS data is located.}

\item{resolution_um}{The image pixel size un micrometers (sorry it can not be read directly from XMASS).}

\item{xml_file_full_path}{Full path to the XML file or alternatively an already parsed xml list resulting from ParseBrukerXML() function.}

\item{txt_spectrum_path}{Full path to a bruker spectrum txt file to extract mass axis (optional).}

\item{selected_img}{the image number to select in the XML file containing various classes.}

\item{...}{Extra parameters to .readBrukerXmassImg ( for example max_ff_file_size_MB as max size in MB of each ff file of the ramdisk).}
}
\description{
This function converts MSI data contained in Bruker XMASS directory into  rMSI data structure.
Only spectra contained in the XML file are imported. The ramdisk will be created at the specified
directory as a collection of ff files with the maximum specified size (50 MB is the default setting).
The rMSI object pointing to ramdisk and containing RAM stored variables is returned.
}
