% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MergeMassAxis}
\alias{MergeMassAxis}
\title{MergeMassAxis.}
\usage{
MergeMassAxis(mz1, mz2)
}
\arguments{
\item{mz1}{the first mass axis to merge.}

\item{mz2}{the second mass axis to merge.}
}
\value{
a list containing the common mass axis that represents mz1 and mz1 accurately and a boolean indicating if and error was raised.
}
\description{
Merges two mass axis in a single one using an apropiate bin size.
The resulting mass axis will display a bin size equal to the minimum of two supplied vectors. 
The bin size is calculated relative to the m/z for better accuracy.
The resulting mass axis range is calculated using the common range between the two mass axis.
If there is no overlao between the two mass axis range an error will be raised.
}
