% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIWindow.R
\name{MSIWindow}
\alias{MSIWindow}
\title{Open the GUI to explore a MS image}
\usage{
MSIWindow(img)
}
\arguments{
\item{img}{a rMSI data object

 Open the GUI to explore the MS image provided as parameter.}
}
\description{
Open the GUI to explore a MS image
}

