% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportBrukerXMASSImage.R
\name{ParseBrukerXML}
\alias{ParseBrukerXML}
\title{ParseBrukerXML.}
\usage{
ParseBrukerXML(xml_path, sel_img = 1, roi_name_only = F)
}
\arguments{
\item{xml_path}{the full path where XML file is stored.}

\item{sel_img}{the image/roi to select in the XML file.}

\item{roi_name_only}{if true, no spots are processed and only the name of the img/roi is returned.}
}
\value{
all spots strings arranged in a list.
}
\description{
Reads a Bruker's xml file exported using fleximaging.
Each Bruker's ROI is stored independently.
A list is returned where each element in the list contains the region name and the spots strings.
Only the region selected by sel_img parameter is parsed each time this function is executed.
To get all regions in a XML file this function must be called for each region.
The total number of regions in an XML file can be determined usin the CountImagesInBrukerXml function.
}
