% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{PlotTICImage}
\alias{PlotTICImage}
\title{PlotTICImage.}
\usage{
PlotTICImage(img, TICs = NULL, rotate = 0, scale_title = "TIC")
}
\arguments{
\item{img}{an rMSI object.}

\item{TICs}{a vector of TIC values ordered acording pos array in img object.}
}
\description{
PlotTICImage.
}
