% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{PlotValues}
\alias{PlotValues}
\title{PlotValues.}
\usage{
PlotValues(posMat, values, rotate = 0, scale_title = "",
  pixel_size_um = 100)
}
\arguments{
\item{posMat}{a two columns matrix where first column ara the x coodrinates of values and second column the y coordinates.}

\item{values}{the values to plot.}

\item{rotate}{rotation to apply.}

\item{scale_title}{a text label for the color scale.}

\item{pixel_size_um}{the pixel resolution in um.}
}
\description{
Plot values in a image using the same methods as plotting an MS image.
The raster position of each value is definied in posMat.
}
