% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imzMLreader.R
\name{xmlParseSpectrum}
\alias{xmlParseSpectrum}
\title{xmlParseSpectrum.}
\usage{
xmlParseSpectrum(xmlSingleSpectrum)
}
\arguments{
\item{xmlSingleSpectrum}{a xml node containing data for a single spectrum.}
}
\value{
a named list containing all parameters of the spectrum.
}
\description{
Parses an XML node in imzML format containing a single spectrum. All meaning data is returned as named list.
Internal use only.
}
